correct some some casterrors

+++ src/graphrep/HashNN.java
@@ -23,8 +23,8 @@ public class HashNN implements NNSearcher {
 		dumpNN = new DumbNN(graphRep);
 		hashMap = new HashMap<Long, Object>();
 		for (int i = 0; i < graphRep.getNodeCount(); i++) {
-			int tempLat = (int) (graphRep.getNodeLat(i) * 1000);
-			int tempLon = (int) (graphRep.getNodeLon(i) * 1000);
+			long tempLat = (long) graphRep.getNodeLat(i) * 1000;
+			long tempLon = (long) graphRep.getNodeLon(i) * 1000;
 
 			long key = tempLat << 32 | tempLon;
 			IntArrayList tempValues = (IntArrayList) hashMap.get(key);
@@ -49,8 +49,8 @@ public class HashNN implements NNSearcher {
 	 */
 	@Override
 	public int getIDForCoordinates(double lat, double lon) {
-		int keyLat = (int) (lat * 1000);
-		int keyLon = (int) (lon * 1000);
+		long keyLat = (long) (lat * 1000);
+		long keyLon = (long) (lon * 1000);
 		int pos = -1;
 		long key;
 		double dist = Long.MAX_VALUE;
